auth: Introduce extra fields to auth cookie.
If the initial auth passes, the calling application has the option
to insert extra fields into the auth cookie that will be validated
on each request.

Also add a helper to extract the extra field from a request.

This way the application can store session-specific data such as
an id, and use that id in further requests to identify the client.