auth: Reuse gen_secret to get the auth token.
Only problem is now the Random state is reset each time
authentication is done. While normally not an issue, there may
be situations where you do not really want this to occur, eg
due to overhead from querying the system for entropy, or for
a a need for reproducible runs.